rename to opacity-stream
